requirements.txt
pyTelegramBotAPI
psutil
python-dotenv
matplotlib
google-generativeai
requests
